Experiment with rest API
This is an attempt to create another Haskell-based HTTP library,
providing Servant-like API introspection abilities without the type
magic.

I like Servant for the possibilities it gives to interpret API
definitions for documentation generation and code generation. I also
like the idea of using plain Haskell functions to implement APIs.

Using Servant requires a lot of type knowledge. Additionally error
messages when your endpoint implementation doesn't match the type aren't
great. This doesn't make Servant a great entrypoint for Haskell
beginners.

I don't want to choose between the possibilities Servant brings and
beginner friendliness. This library tries to offer both by defining an
API not as a type but as a GADT, that can be interpreted in different
ways.

This commit offers an incomplete proof of concept that only deals with
the request path.