Interpret the API as a WAI application
By interpreting the API as a WAI application we make the use of Scotty
an implementation detail. Conceivably there might come a point at which
we want to remove the dependency on Scotty and implement the logic of
parsing the request ourselves. Now we can do that without changing the
API of the library.