Allow custom response status and encoders
We weren't able to set the status code or headers of responses before,
but now we can.

I'm a bit concerned that by being hidden behind a function call, the
headers and statuses a certain value can set aren't available to
API introspection. Any response might theoretically set any header
and/or response code. That might be something to improve upon.