Replace `ioApplication` with `sandbox`
I created `ioApplication` with the idea that it would be an easier
entrypoint into this library. At second though I'm not sure if building
an application based on arbitrary IO is ever a good idea.

So let's take a leaf out of Elm's book and make a non-side-effecty
program the entry program.