Use clearer name than 'Identity' for sandbox apps
We're not really using `Identity` for its values at this point, just to
get nice types. Also, someone might want to use `Identity` only to be
confused when its behavior is special.

For new users starting out with this library by learning to create a
`sandbox` application `NoEffects` is going to be clearer than
`Identity`. It's also going to train them to see this type variable as
describing the type of effects the application allows.