Add support for other request methods
There's a couple of methods I'm not yet sure what to do with:

`HEAD` should return the same data as `GET`, just no response body.
Maybe we should generate it automatically based on the `GET` request?
Scotty might already be doing that, it doesn't expose a function for
adding a `head` endpoint itself.

`OPTIONS` sounds like something we should be able to generate from what
we can introspect on the UI.