Allow setting constant response headers
Of the different headers I might set on a response I can think of, a
number of them don't depend on the incoming request in any way. Things
like caching policies are going to be the constant.

This allows defining a header in the endpoint definition, so you don't
need to deal with it in your handler function.